Refactoring codecs
