+++ package.json
@@ -59,14 +59,12 @@
     "webpack-bundle-analyzer": "^2.13.1",
     "webpack-cli": "^2.1.5",
     "webpack-dev-server": "^3.1.5",
-    "webpack-plugin-replace": "^1.1.1"
-  },
-  "dependencies": {
+    "webpack-plugin-replace": "^1.1.1",
     "classnames": "^2.2.6",
     "comlink": "^3.0.3",
-    "comlink-loader": "^1.0.0",
-    "preact": "^8.3.1",
     "linkstate": "^1.1.1",
-    "pretty-bytes": "^5.1.0"
+    "preact": "^8.3.1",
+    "pretty-bytes": "^5.1.0",
+    "worker-plugin": "^1.1.1"
   }
 }
