+++ src/codecs/browser-jp2/encoder.ts
@@ -1,15 +1,5 @@
+import { EncodeOptions, mimeType } from './encoder-meta';
 import { canvasEncode } from '../../lib/util';
-import { canvasEncodeTest } from '../generic/util';
-
-export interface EncodeOptions { }
-export interface EncoderState { type: typeof type; options: EncodeOptions; }
-
-export const type = 'browser-jp2';
-export const label = 'Browser JPEG 2000';
-export const mimeType = 'image/jp2';
-export const extension = 'jp2';
-export const defaultOptions: EncodeOptions = {};
-export const featureTest = () => canvasEncodeTest(mimeType);
 
 export function encode(data: ImageData, options: EncodeOptions) {
   return canvasEncode(data, mimeType);
