+++ src/codecs/browser-pdf/encoder.ts
@@ -1,15 +1,5 @@
+import { EncodeOptions, mimeType } from './encoder-meta';
 import { canvasEncode } from '../../lib/util';
-import { canvasEncodeTest } from '../generic/util';
-
-export interface EncodeOptions { }
-export interface EncoderState { type: typeof type; options: EncodeOptions; }
-
-export const type = 'browser-pdf';
-export const label = 'Browser PDF';
-export const mimeType = 'application/pdf';
-export const extension = 'pdf';
-export const defaultOptions: EncodeOptions = {};
-export const featureTest = () => canvasEncodeTest(mimeType);
 
 export function encode(data: ImageData, options: EncodeOptions) {
   return canvasEncode(data, mimeType);
