+++ src/codecs/resize/options.tsx
@@ -2,7 +2,7 @@ import { h, Component } from 'preact';
 import linkState from 'linkstate';
 import { bind } from '../../lib/initial-util';
 import { inputFieldValueAsNumber } from '../../lib/util';
-import { ResizeOptions } from './resize';
+import { ResizeOptions } from './processor-meta';
 
 interface Props {
   isVector: Boolean;
