+++ src/missing-types.d.ts
@@ -22,3 +22,19 @@ declare module '*.svg' {
   const content: string;
   export default content;
 }
+
+declare module '*.wasm' {
+  const content: string;
+  export default content;
+}
+
+interface WorkerOptions {
+  type?: 'classic' | 'module';
+}
+
+interface Window {
+  Worker: {
+    prototype: Worker;
+    new(stringUrl: string, options?: WorkerOptions): Worker;
+  };
+}
